requirements.txt
pytesseract
pillow
mss
tk
obs-websocket-py
opencv-python
numpy
timm
torch
transformers
